Removed for now an extended property for Collabora as it actually breaks
The observed behavior is that, in a collaborative session, if a user
goes out and back in it might trigger a "Document changed" message for everyone,
leaving the document in a locked state. Or Collabora may get into a loop
attempting to reload the "latest revision" over and over again. To be followed up.