Hashed the inode obtained by the cs3 storage interface
This is because applications iteratively decode urlencoded inodes,
eventually making WOPI calls fail. This change ensures a numeric
inode even when Reva runs with the local storage provider.
As a consequence, it is not any longer possible to "reverse" stat
by inode with cs3, thus the related test has been removed.