Improved logging + fixes for the islock=True writes
Notably the fix for xrootd/eos is to NOT use the atomic flag
when writing a file in O_EXCL mode. This makes sure that
an error is caught right at open time. Alternatively, at close
time an atomic concurrent write would still fail with:
"file has been cleaned because of an overlapping atomic upload"
(which was not caught).