Moved interop locking logic to storeWopiLock
Instead of early checking for existing MS or LO locks, the logic
is now moved to storeWopiLock, which fails if external locks
are present. This is safer and allows the UI to control the
interaction with the user, instead of injecting a "(locked by
another app)" message in the BreadcrumbDocName.