Use the extended WOPI Lock with validation when relocking files
This helps detecting externally modified files when a session was
put on standby (e.g. by closing the lid) and resumed after: as the
lock was removed, to protect from overwriting potentially new content,
the WOPI server will check if the file was last saved in WOPI or not,
and in the first case will refresh the lock even without a previously
existing lock (as required by the RefreshLock API).