Changed strategy with lock mismatches on PutFile
We have seen false positives where a PutFile with a mismatched lock is
followed by the correct one, therefore it is better not to pollute the
user's spaces with worrying conflict files. However, we have seen true
positives as well, as being discussed at:

https://www.yammer.com/officeenterprisenda/#/Threads/show?threadId=2334861341941760

Therefore we want to keep a copy for later recovery.